requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1

# LangChain + Google Gemini
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0

# Vector store
chromadb>=0.5.0

# Loaders & utils
pypdf>=4.0.0
python-multipart>=0.0.9

# Testing
pytest>=8.2.0
httpx>=0.27.0
requests>=2.28.0
motor>=3.1.1
